requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
numpy
python-multipart
pydantic